% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_est_pwK.R
\name{gamma_fun}
\alias{gamma_fun}
\title{Estimate proportional cause ratio model parameters}
\usage{
gamma_fun(data, ind.wt)
}
\arguments{
\item{data}{Clinical trial dataset. The data should contain \code{Y.R}, \code{Delta.R},
\code{delta.R}, \code{X1.R}, and \code{X2.R}. If \code{ind.wt = 1}, the data should also contain
a perturbation weight variable named \code{rwt}.}

\item{ind.wt}{Indicator for perturbation weighting. Use \code{0} for the original
estimation and \code{1} for weighted estimation.}
}
\value{
A list with components:
\describe{
\item{coeff}{Estimated proportional cause ratio model coefficients.}
\item{coeff_se}{Estimated standard errors of the coefficients.}
\item{vcov}{Estimated variance-covariance matrix.}
}
}
\description{
Estimates the parameters in the proportional cause ratio model using clinical
trial data with observed cause-of-death information.
}
