% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_est_pwK.R
\name{neglog_pl_beta}
\alias{neglog_pl_beta}
\title{Negative log partial likelihood for beta estimation}
\usage{
neglog_pl_beta(odat, beta, gammas.hat, events_id, ind.wt)
}
\arguments{
\item{odat}{Observational study dataset.}

\item{beta}{Numeric vector of disease-specific survival regression coefficients.}

\item{gammas.hat}{Estimated proportional cause ratio model parameters.}

\item{events_id}{Integer vector indicating rows corresponding to observed events.}

\item{ind.wt}{Indicator for perturbation weighting. Use \code{0} for the original
estimation and \code{1} for weighted estimation.}
}
\value{
A numeric value of the negative log partial likelihood.
}
\description{
Internal objective function used in nonlinear optimization for estimating
disease-specific survival regression coefficients.
}
\keyword{internal}
